+++ NEWS.md
@@ -1,6 +1,7 @@
 ingredients 0.3.2
 ----------------------------------------------------------------
 * added new functions `describe()` and `print.ceteris_paribus_descriptions()` for text based descriptions of Ceteris Paribus explainers
+* `plot.ceteris_paribus_explainer` works now also for categorical variables. Use the `only_numerical = FALSE` to force bars
 
 ingredients 0.3.1
 ----------------------------------------------------------------